Decoupling tests
- Split the PIRLinker and PIRIngestor tests
- Update test_data_2008 to be compatible with changes
to handling of section in merge_question_response.
